backend/pyproject.toml
[tool.poetry]
name = "vaaniai-backend"
version = "0.1.0"
description = "VaaniAI — AI Voice Translation & Conversation Platform Backend"
authors = ["VaaniAI Team"]
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = { extras = ["standard"], version = "^0.30.0" }
pydantic = "^2.7.0"
pydantic-settings = "^2.3.0"
sqlalchemy = "^2.0.30"
alembic = "^1.13.0"
asyncpg = "^0.29.0"
psycopg2-binary = "^2.9.9"
redis = "^5.0.4"
python-jose = { extras = ["cryptography"], version = "^3.3.0" }
passlib = { extras = ["bcrypt"], version = "^1.7.4" }
python-multipart = "^0.0.9"
httpx = "^0.27.0"
openai = "^1.30.0"
faster-whisper = "^1.0.2"
langchain = "^0.2.0"
langchain-openai = "^0.1.8"
langchain-pinecone = "^0.1.1"
pinecone-client = "^3.2.2"
stripe = "^9.12.0"
fasttext-wheel = "^0.9.2"
boto3 = "^1.34.0"
aiofiles = "^23.2.1"
celery = { extras = ["redis"], version = "^5.4.0" }
flower = "^2.0.1"
websockets = "^12.0"
numpy = "^1.26.0"
soundfile = "^0.12.1"
librosa = "^0.10.2"
transformers = "^4.41.0"
torch = "^2.3.0"
sentencepiece = "^0.2.0"
sacremoses = "^0.1.1"
indic-nlp-library = "^0.92"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
pytest-asyncio = "^0.23.7"
pytest-cov = "^5.0.0"
httpx = "^0.27.0"
ruff = "^0.4.7"
mypy = "^1.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
